Revert "Refactor calculator routes into a Next route group"
This reverts commit 950dcaa4cb6d7d2d71c37998817be0698584bd80.
