+++ app/src/libs/calculations/CalcOrchestratorManager.ts
@@ -71,8 +71,7 @@ export class CalcOrchestratorManager {
     } catch (error) {
       console.error(`[CalcOrchestratorManager] Failed to start calculation ${calcId}:`, error);
       this.cleanup(calcId);
-      // Avoid bubbling async startup errors into the Next.js router.
-      // The error is logged and the orchestrator is cleaned up.
+      throw error;
     }
   }
