+++ app/src/pathways/population/PopulationPathwayWrapper.tsx
@@ -7,6 +7,7 @@
 
 import { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
+import StandardLayout from '@/components/StandardLayout';
 import { CURRENT_YEAR } from '@/constants';
 import { useAppNavigate } from '@/contexts/NavigationContext';
 import { ReportYearProvider } from '@/contexts/ReportYearContext';
@@ -143,5 +144,9 @@ export default function PopulationPathwayWrapper({ onComplete }: PopulationPathw
       currentView = <></>;
   }
 
-  return <ReportYearProvider year={CURRENT_YEAR}>{currentView}</ReportYearProvider>;
+  return (
+    <ReportYearProvider year={CURRENT_YEAR}>
+      <StandardLayout>{currentView}</StandardLayout>
+    </ReportYearProvider>
+  );
 }
