--- calculator-app/src/app/[countryId]/(calculator)/households/create/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import PopulationPathwayWrapper from "@/pathways/population/PopulationPathwayWrapper";
-
-export default function HouseholdCreateRoute() {
-  return <PopulationPathwayWrapper />;
-}
