--- calculator-app/src/app/[countryId]/(calculator)/households/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import PopulationsPage from "@/pages/Populations.page";
-
-export default function HouseholdsRoute() {
-  return <PopulationsPage />;
-}
