--- calculator-app/src/app/[countryId]/(calculator)/policies/create/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import PolicyPathwayWrapper from "@/pathways/policy/PolicyPathwayWrapper";
-
-export default function PolicyCreateRoute() {
-  return <PolicyPathwayWrapper />;
-}
