--- calculator-app/src/app/[countryId]/(calculator)/policies/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import PoliciesPage from "@/pages/Policies.page";
-
-export default function PoliciesRoute() {
-  return <PoliciesPage />;
-}
