--- calculator-app/src/app/[countryId]/(calculator)/reports/create/[userReportId]/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import { use } from "react";
-import ModifyReportPage from "@/pages/reportBuilder/ModifyReportPage";
-
-export default function ModifyReportRoute({
-  params,
-}: {
-  params: Promise<{ userReportId: string }>;
-}) {
-  const { userReportId } = use(params);
-
-  return <ModifyReportPage userReportId={userReportId} />;
-}
