--- calculator-app/src/app/[countryId]/(calculator)/reports/create/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import ReportBuilderPage from "@/pages/reportBuilder/ReportBuilderPage";
-
-export default function ReportBuilderRoute() {
-  return <ReportBuilderPage />;
-}
