--- calculator-app/src/app/[countryId]/(calculator)/reports/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import ReportsPage from "@/pages/Reports.page";
-
-export default function ReportsRoute() {
-  return <ReportsPage />;
-}
