--- calculator-app/src/app/[countryId]/(calculator)/simulations/create/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import SimulationPathwayWrapper from "@/pathways/simulation/SimulationPathwayWrapper";
-
-export default function SimulationCreateRoute() {
-  return <SimulationPathwayWrapper />;
-}
