--- calculator-app/src/app/[countryId]/(calculator)/simulations/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import SimulationsPage from "@/pages/Simulations.page";
-
-export default function SimulationsRoute() {
-  return <SimulationsPage />;
-}
