+++ calculator-app/src/app/[countryId]/households/create/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import PopulationPathwayWrapper from "@/pathways/population/PopulationPathwayWrapper";
+import { CalculatorProviders } from "../../providers";
+
+export default function HouseholdCreateRoute() {
+  return (
+    <CalculatorProviders>
+      <PopulationPathwayWrapper />
+    </CalculatorProviders>
+  );
+}
