+++ calculator-app/src/app/[countryId]/households/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import StandardLayout from "@/components/StandardLayout";
+import PopulationsPage from "@/pages/Populations.page";
+import { CalculatorProviders } from "../providers";
+
+export default function HouseholdsRoute() {
+  return (
+    <CalculatorProviders>
+      <StandardLayout>
+        <PopulationsPage />
+      </StandardLayout>
+    </CalculatorProviders>
+  );
+}
