+++ calculator-app/src/app/[countryId]/policies/create/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import PolicyPathwayWrapper from "@/pathways/policy/PolicyPathwayWrapper";
+import { CalculatorProviders } from "../../providers";
+
+export default function PolicyCreateRoute() {
+  return (
+    <CalculatorProviders>
+      <PolicyPathwayWrapper />
+    </CalculatorProviders>
+  );
+}
