+++ calculator-app/src/app/[countryId]/report-output/[reportId]/[[...rest]]/page.tsx
@@ -1,7 +1,9 @@
 "use client";
 
 import { use } from "react";
+import StandardLayout from "@/components/StandardLayout";
 import ReportOutputPage from "@/pages/ReportOutput.page";
+import { CalculatorProviders } from "../../../providers";
 
 export default function ReportOutputRoute({
   params,
@@ -12,5 +14,11 @@ export default function ReportOutputRoute({
   const subpage = rest?.[0];
   const view = rest?.[1];
 
-  return <ReportOutputPage reportId={reportId} subpage={subpage} view={view} />;
+  return (
+    <CalculatorProviders>
+      <StandardLayout>
+        <ReportOutputPage reportId={reportId} subpage={subpage} view={view} />
+      </StandardLayout>
+    </CalculatorProviders>
+  );
 }
