+++ calculator-app/src/app/[countryId]/reports/create/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import StandardLayout from "@/components/StandardLayout";
+import ReportBuilderPage from "@/pages/reportBuilder/ReportBuilderPage";
+import { CalculatorProviders } from "../../providers";
+
+export default function ReportBuilderRoute() {
+  return (
+    <CalculatorProviders>
+      <StandardLayout>
+        <ReportBuilderPage />
+      </StandardLayout>
+    </CalculatorProviders>
+  );
+}
