+++ calculator-app/src/app/[countryId]/reports/page.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import StandardLayout from "@/components/StandardLayout";
+import ReportsPage from "@/pages/Reports.page";
+import { CalculatorProviders } from "../providers";
+
+export default function ReportsRoute() {
+  return (
+    <CalculatorProviders>
+      <StandardLayout>
+        <ReportsPage />
+      </StandardLayout>
+    </CalculatorProviders>
+  );
+}
