+++ calculator-app/src/app/[countryId]/simulations/create/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import SimulationPathwayWrapper from "@/pathways/simulation/SimulationPathwayWrapper";
+import { CalculatorProviders } from "../../providers";
+
+export default function SimulationCreateRoute() {
+  return (
+    <CalculatorProviders>
+      <SimulationPathwayWrapper />
+    </CalculatorProviders>
+  );
+}
